wslc: align Size/ShmSize types with docker int64 end-to-end
Per PR review: replace boundary static_casts with consistent signed
64-bit types across IDL, SDK, service models, and CLI schema. Mirrors
the existing MemoryBytes/NanoCpus precedent.

- IDL: WSLCImageInformation.Size, WSLCContainerOptions.ShmSize -> LONGLONG
- SDK: WslcImageInfo.sizeBytes -> int64_t
- Models: ImageInformation::Size, ContainerOptions::ShmSize -> int64_t
- Schema: InspectImage::Size -> int64_t
- Launcher: m_shmSize/SetShmSize -> int64_t
- Validation: GetMemorySizeFromString returns int64_t

+++ src/windows/wslc/services/ImageService.cpp
@@ -204,7 +204,7 @@ std::vector<ImageInformation> ImageService::List(wsl::windows::wslc::models::Ses
 
         info.Id = image.Hash;
         info.Created = image.Created;
-        info.Size = static_cast<uint64_t>(image.Size);
+        info.Size = image.Size;
         result.push_back(info);
     }
